% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{pkgdown_template}
\alias{pkgdown_template}
\title{State- and University Library of Göttingen \href{http://pkgdown.r-lib.org}{pkgdown} template}
\usage{
pkgdown_template()
}
\value{
named list of paths to assets and templates for web design
}
\description{
State- and University Library of Göttingen \href{http://pkgdown.r-lib.org}{pkgdown} template
}
\section{Pkgdown Usage}{

To use this template for your pkgdown website, list subugoetheme in your \code{DESCRIPTION} and add to your \verb{_pkgdown.yml}:\if{html}{\out{<div class="yaml">}}\preformatted{template:
  package: subugoetheme
}\if{html}{\out{</div>}}

To update your website, install the current version of {subugoetheme} and re-run \code{pkgdown::build_site()} locally or in your CI.
\subsection{Features:}{

As a result, your website will:
\itemize{
\item Use the \href{https://bootswatch.com/3/cosmo/}{\strong{bootswatch \code{cosmo} bootstrap theme}} with an inverse navbar.
The corresponding settings in you \verb{_pkgdown.yml} (\code{template: params: bootswatch}, \code{navbar: type: inverse}) are ignored.
\item Display in a University of Göttingen \strong{color scheme}.
This is determined by the \code{ugoe_bootswatch.css}, which is compiled from \code{ugoe_colors.less}.
The colors were added and downloaded via https://www.bootstrap-live-customizer.com.
\item Include the \strong{SUB wordmark} in your website as \code{sub_wordmark.jpg}.
The wordmark is included in the custom footer.
\item Expose the SUB twitter account \href{https://twitter.com/subugoe}{\verb{@subugoe}} as \strong{opengraph \href{https://pkgdown.r-lib.org/articles/metadata.html}{metadata}} (\code{twitter: site}).
\item Link to the SUB legal boilerplate.
}

For additional features, see \href{http://subugoe.github.io/muggle/}{muggle}.
}

\subsection{Analytics:}{

Instead of Google Analytics, this template includes a tracking code for \code{matomo.gwdg.de} on every website.
Matomo is an open-source alternative to Google Analytics, here hosted by the \href{http://matomo.gwdg.de}{GWDG}.

To opt out, users are directed to the matomo opt-out at sub.uni-goettingen.de.
The gwdg matomo instance respects the opt-out set on sub.uni-goettingen.de.

The following SUB employees have access to the analytics:
\itemize{
\item \href{https://www.sub.uni-goettingen.de/kontakt/personen-a-z/personendetails/person/najko-jahn/}{Najko Jahn}
\item \href{https://www.sub.uni-goettingen.de/kontakt/personen-a-z/personendetails/person/maximilian-held/}{Maximilian Held}
\item \code{scholcomm-analytics@sub.uni-goettingen.de}, a function account for \href{https://www.sub.uni-goettingen.de/kontakt/abteilungen-a-z/abteilungs-und-gruppendetails/abteilunggruppe/wissen-als-gemeingut/}{WAG} data products.
}
}

\subsection{Maintenance:}{

To pull in upstream changes from {pkgdown} into \emph{this} package, it may be ocassionally neccessary to update the files in \code{inst/pkgdown/templates} which have been minimally adapted from their original counterparts in {pkgdown}.

To learn more about pkgdown templates, read \href{https://pkgdown.r-lib.org/reference/build_site.html}{here} or consider \href{https://github.com/ropensci/rotemplate}{these} \href{https://tidytemplate.tidyverse.org}{examples}.
}
}

\section{Other Usage}{

The CSS targets bootstrap 3.x and can thus be used on other bootstrap 3.x websites.
\subsection{flexdashboard}{

To use the CSS for \href{https://rmarkdown.rstudio.com/flexdashboard/}{flexdasboard}, list subugoetheme in your \code{DESCRIPTION} and add to your yaml frontmatter:\preformatted{output:
 flexdashboard::flex_dashboard:
   css: !expr subugoetheme::pkgdown_template()$assets$css
}
}
}

\concept{corporate identity}
